+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 - Center Crosshair (like it was in older versions around <=1.12.2)
 - Old Cloud Rendering (<=1.21.5)
 - Old Water Overlay Opacity (the texture when you are submerged) (<=1.12.2)
+- Old Block Overlay Opacity (the texture when you are inside a block) (<=1.12.2)
 
 ## Changed:
 #### General